# Fix scene performance profiling test